+++ modules/Material/TextField.qml
@@ -1,6 +1,7 @@
 /*
  * QML Material - An application framework implementing Material Design.
  * Copyright (C) 2014 Bogdan Cuza <bogdan.cuza@hotmail.com>
+ *               2015 Ricardo Vieira <ricardo.vieira@tecnico.ulisboa.pt>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -15,205 +16,20 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import QtQuick 2.2
+
+import QtQuick 2.4
+import QtQuick.Controls 1.3 as Controls
+import QtQuick.Controls.Styles.Material 0.1 as MaterialStyle
 import Material 0.1
 
-/*
- * Note that this is a FocusScope, not a TextInput. If you need to read or
- * write properties of the input itself, use the input property.
- */
-FocusScope {
-   id: field
+Controls.TextField {
 
-   property alias acceptableInput: textInput.acceptableInput
-   property alias validator: textInput.validator
-   property alias inputMask: textInput.inputMask
+    property color color: Theme.primaryColor
+    property color errorColor: Palette.colors["red"]["500"]
+    property string helperText
+    property bool floatingLabel: false
+    property bool hasError: characterLimit && length > characterLimit
+    property int characterLimit
 
-   property color accentColor: Theme.accentColor
-   property color errorColor: "#F44336"
-
-   property alias text: textInput.text
-
-   property alias placeholderText: fieldPlaceholder.text
-   property alias helperText: helperTextLabel.text
-
-   readonly property int characterCount: text.length
-
-   property bool floatingLabel: false
-   property bool hasError: false
-   property int characterLimit
-
-   readonly property rect inputRect: Qt.rect(textInput.x, textInput.y, textInput.width, textInput.height)
-
-   readonly property alias input: textInput
-
-   signal accepted()
-   signal editingFinished()
-
-   implicitHeight: __internal.showHelperText ? helperTextLabel.y + helperTextLabel.height + units.dp(4)
-                          : underline.y + units.dp(8)
-   width: units.dp(200)
-
-   QtObject {
-       id: __internal
-
-       property bool showHelperText: helperText.length > 0
-       property bool showCharacterCounter: characterLimit > 0
-   }
-
-   MouseArea {
-      anchors.fill: parent
-      onClicked: field.forceActiveFocus(Qt.MouseFocusReason)
-   }
-
-   TextInput {
-      id: textInput
-
-      focus: true
-
-      color: Theme.light.textColor
-      selectedTextColor: "white"
-      selectionColor: Qt.darker(field.accentColor, 1)
-
-      selectByMouse: Device.type === Device.desktop
-
-      activeFocusOnTab: true
-
-      width: parent.width
-
-      clip: true
-
-      y: {
-         if( !floatingLabel )
-            return units.dp(16)
-         if( floatingLabel && !__internal.showHelperText )
-            return units.dp(40)
-         return units.dp(28)
-      }
-
-      font {
-         family: echoMode == TextInput.Password && field.text.length > 0 ? "" : "Roboto"
-         pixelSize: units.dp(16)
-      }
-
-      onAccepted: field.accepted()
-      onEditingFinished: field.editingFinished()
-
-      MouseArea {
-          anchors.fill: parent
-          cursorShape: Qt.IBeamCursor
-          acceptedButtons: Qt.NoButton
-      }
-   }
-
-   Label {
-      id: fieldPlaceholder
-
-      text: field.placeholderText
-
-      font.pixelSize: units.dp(16)
-
-      anchors.baseline: textInput.baseline
-      anchors.bottomMargin: units.dp(8)
-
-      color: Theme.light.hintColor
-
-      states: [
-         State {
-            name: "floating"
-            when: textInput.displayText.length > 0 && floatingLabel
-            AnchorChanges {
-               target: fieldPlaceholder
-               anchors.baseline: undefined
-               anchors.bottom: textInput.top
-            }
-            PropertyChanges {
-               target: fieldPlaceholder
-               font.pixelSize: units.dp(12)
-            }
-         },
-         State {
-            name: "hidden"
-            when: textInput.displayText.length > 0 && !floatingLabel
-            PropertyChanges {
-               target: fieldPlaceholder
-               visible: false
-            }
-         }
-      ]
-
-      transitions: [
-         Transition {
-            id: floatingTransition
-            enabled: false
-            AnchorAnimation {
-               duration: 200
-            }
-            NumberAnimation {
-               duration: 200
-               property: "font.pixelSize"
-            }
-         }
-      ]
-
-      Component.onCompleted: floatingTransition.enabled = true
-   }
-
-   Rectangle {
-      id: underline
-      color: field.hasError || (__internal.showCharacterCounter && field.characterCount > field.characterLimit)
-             ? field.errorColor : field.activeFocus ? field.accentColor : Theme.light.hintColor
-
-      height: field.activeFocus ? units.dp(2) : units.dp(1)
-
-      anchors {
-         left: parent.left
-         right: parent.right
-         top: textInput.bottom
-         topMargin: units.dp(8)
-      }
-
-      Behavior on height {
-         NumberAnimation { duration: 200 }
-      }
-
-      Behavior on color {
-         ColorAnimation { duration: 200 }
-      }
-   }
-
-   Label {
-      id: helperTextLabel
-      visible: __internal.showHelperText
-      font.pixelSize: units.dp(12)
-      color: field.hasError ? field.errorColor : Qt.darker(Theme.light.hintColor)
-      anchors {
-         left: parent.left
-         right: parent.right
-         top: underline.top
-         topMargin: units.dp(4)
-      }
-
-      Behavior on color {
-         ColorAnimation { duration: 200 }
-      }
-   }
-
-   Label {
-      id: characterCounterLabel
-      visible: __internal.showCharacterCounter
-      font.pixelSize: units.dp(12)
-      font.weight: Font.Light
-      color: field.characterCount <= field.characterLimit ? Qt.darker(Theme.light.hintColor) : field.errorColor
-      text: field.characterCount + " / " + field.characterLimit
-      anchors {
-         right: parent.right
-         top: underline.top
-         topMargin: units.dp(8)
-      }
-
-      Behavior on color {
-         ColorAnimation { duration: 200 }
-      }
-   }
+    style: MaterialStyle.TextFieldStyle {}
 }